pyproject.toml
[project]
name = "litebot-new"
version = "0.1.0"
description = "litebot-new"
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "aiohttp>=3.12.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-mclib>=0.0.3.post2",
    "nonebot-plugin-tortoise-orm>=0.1.4",
    "nonebot2[fastapi]>=2.4.2",
    "ping3>=4.0.8",
    "pip>=25.1.1",
    "psutil>=7.0.0",
    "pydantic>=2.11.5",
    "python-nmap>=0.7.1",
    "qrcode>=8.2",
    "tortoise-orm[accel,asyncmy]>=0.25.0",
    "watchdog>=6.0.0",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = [
    "nonebot_plugin_tortoise_orm",
    "nonebot_plugin_mclib",
    "nonebot_plugin_localstore",
]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
    "I",     # isort
    "PERF",  # pylint-performance
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "E501",   # line-too-long
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.pyright]
typeCheckingMode = "basic"

[dependency-groups]
dev = []
